website: [1025] Removal of both "referring" and "download" statistics, untangled Gold Star nominations.
Between Google Analytics and Urchin we don't ned these statistics anymore, and they just muck up the code and slow things down.  (I anticipate a speed increase with the website when this is deployed.)

Also untangled the gold star nominations from the rest of it.  The design obviously regarded them as statistics, but they aren't used as such which made their intermingling with the other statistics weird and difficult to use.

Updated all references to these files, removed files and sections of code that is/are no longer relevant.

Note to self: nominate-utils.php has a naming convention that is inconsistent with the style in the rest of the website.  This has not been addressed in this checkin.