website: Semi-major change to the way that the "referrer" and the referring page is handled.
It was often being confused with a directive telling a page where it should go to next after a user action (such as submitting a contribution or deleting a comment).  It was extremely unweildy, wildy overused, and delicate.  This change is a change to the way it is handled.  Much of the "referrer" stuff has been removed entirely, what was left usually changed names to make the intent more clear.  There are still some cases where the actual referrer is overridden with a fake referrer and used to direct the page, hopefully these cases are more understandable.  In the future it would be better to have a more generalized framework for handling this, but that is beyond the scope of this change.