Updated the licensing on the website, see #1566.
The website licensing is a seperate file and is included into the
licensing page, rather than having everything in one place.  Updating
the licensing should update the respective license file.

The License_website_v7.htm file was modified from the original given
in the ticket to get it to render properly with the PhET website.
When updating, similar processing may be necessary.